fixes for windows

+++ sdk/core/performance-stress/inc/azure/performance-stress/test_options.hpp
@@ -14,6 +14,6 @@ namespace Azure { namespace PerformanceStress {
     std::string Name;
     std::vector<std::string> Activators;
     std::string DisplayMessage;
-    uint expectedArgs;
+    uint16_t expectedArgs;
   };
 }} // namespace Azure::PerformanceStress
